+++ js/data.js
@@ -1,3 +1,5 @@
+import { getRandomInteger, createRandomIdFromRangeGenerator } from './util.js';
+
 const MESSAGES = [
   'Всё отлично!',
   'В целом всё неплохо. Но не всё.',
@@ -17,26 +19,31 @@ const NAMES = [
   'Люпита',
   'Вашингтон',
 ];
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
+const createComment = () => ({
+  id: createRandomIdFromRangeGenerator(1, 1000) ,
+  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
+  message: MESSAGES[getRandomInteger(0, MESSAGES.length - 1)],
+  name: NAMES[getRandomInteger(0, NAMES.length - 1)],
+});
 
-function createRandomIdFromRangeGenerator (min, max) {
-  const previousValues = [];
-
-  return function () {
-    let currentValue = getRandomInteger(min, max);
-    if (previousValues.length >= (max - min + 1)) {
-      return null;
-    }
-    while (previousValues.includes(currentValue)) {
-      currentValue = getRandomInteger(min, max);
-    }
-    previousValues.push(currentValue);
-    return currentValue;
+function createPhoto () {
+  const comments = [];
+  for (let i = 0; i <= getRandomInteger(0, 30); i++) {
+    comments.push(createComment());
+  }
+  return {
+    id: getRandomInteger(1, 25),
+    url: `photos/${getRandomInteger(1, 25)}.jpg`,
+    description: 'фото',
+    likes: createRandomIdFromRangeGenerator(15, 200),
+    comments: comments
   };
 }
-export{MESSAGES, NAMES, getRandomInteger, createRandomIdFromRangeGenerator};
+function createArrayOfPhotos () {
+  const arrayOfPhotos = [];
+  for (let i = 0; i < 25; i++) {
+    arrayOfPhotos.push(createPhoto());
+  }
+}
+
+export { MESSAGES, NAMES, createArrayOfPhotos };
