+++ js/util.js
@@ -1,29 +1,24 @@
-import {MESSAGES, NAMES, getRandomInteger, createRandomIdFromRangeGenerator} from './data.js';
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
 
-const createComment = () => ({
-  id: createRandomIdFromRangeGenerator(1, 1000) ,
-  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-  message: MESSAGES[getRandomInteger(0, MESSAGES.length - 1)],
-  name: NAMES[getRandomInteger(0, NAMES.length - 1)],
-});
+function createRandomIdFromRangeGenerator (min, max) {
+  const previousValues = [];
 
-function createPhoto () {
-  const comments = [];
-  for (let i = 0; i <= getRandomInteger(0, 30); i++) {
-    comments.push(createComment());
-  }
-  return {
-    id: getRandomInteger(1, 25),
-    url: `photos/${getRandomInteger(1, 25)}.jpg`,
-    description: 'фото',
-    likes: createRandomIdFromRangeGenerator(15, 200),
-    comments: comments
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
   };
 }
-function createArrayOfPhotos () {
-  const ArrayOfPhotos = [];
-  for (let i = 0; i < 25; i++) {
-    ArrayOfPhotos.push(createPhoto());
-  }
-}
-export{createArrayOfPhotos};
+
+export { getRandomInteger, createRandomIdFromRangeGenerator };
